Support ReSharper Out-of-Process mode
Currently, the zone activator activates PSI-related zones in Visual
Studio Side unconditionally, which renders ReSharper OOP mode useless.
Code analysis services will performed on both sides in this case.

Disable the activator in Visual Studio when the product is out of
process mode, and add an additional activator for the separate process
mode. Additional activator is reponsible for activation of the extension
in the separate process.